requirements.txt
langchain==0.0.207
flake8==6.0.0
pyright==1.1.316
python-dotenv==1.0.0
colorama==0.4.4
inquirer==3.1.3
Flask==2.3.2
fastapi==0.97.0
uvicorn==0.22.0
openai==0.27.8
prompt-toolkit==3.0.38
tiktoken==0.4.0
google-auth
google-auth-oauthlib
google-cloud-aiplatform
vertexai
requests